fetch editions script on startup
